services/signals/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
mysql-connector-python==8.2.0
pydantic==2.5.0
openai>=1.0.0
requests>=2.28.0
pandas>=1.5.0
numpy>=1.20.0
scikit-learn>=1.3.0
xgboost>=1.7.0
joblib>=1.3.0
python-dotenv>=1.0.0
asyncio
aiofiles
httpx
aiohttp
